Destructure hash in multi-promise-handler
This is a workaround to ensure the identity of the hash changes when its properties are updated. Currently, the properties may change but the hash identity stays the same, so updates don't fully propagate in the latest versions of Ember. It's unclear if this will be the long-term fix, but this solves the issue we have for now until further investigation can be done.